animation

+++ Code/src/UnitePhysic.h
@@ -8,7 +8,7 @@ class UnitePhysic : public Unite
 {
 public:
 	virtual ~UnitePhysic() = default;
-	UnitePhysic(int _x, int _y, int _h, int _l, std::string textureName, std::map<std::string, sf::Texture>& textDictionnary, b2BodyType bodyType, b2World &world);
+	UnitePhysic(float _x, float _y, float _h, float _l, std::string textureName, std::map<std::string, sf::Texture>& textDictionnary, b2BodyType bodyType, b2World &world);
 	b2Body* getBox();
 	void setXY(int _x, int _y) override;
